requirements.txt
dtlpy
opencv-python
numpy
Pillow
